port other hard coded state to their own slices. break out collections to their own component
