+++ src/components/collections-overview/collections-overview.styles.scss
@@ -1,0 +1,4 @@
+.collections-overview {
+  display:flex;
+  flex-direction:column;
+}
