+++ src/components/directory/directory.tsx
@@ -1,44 +1,11 @@
 import * as React from 'react';
 import { MenuItem } from '../menu-item/menu-item';
 import './directory.styles.scss';
-
-export const sectionData = [
-    {
-        title: 'hats',
-        imageUrl: 'https://i.ibb.co/cvpntL1/hats.png',
-        id: 1,
-        linkUrl: 'shop/hats',
-    },
-    {
-        title: 'jackets',
-        imageUrl: 'https://i.ibb.co/px2tCc3/jackets.png',
-        id: 2,
-        linkUrl: 'shop/jackets',
-    },
-    {
-        title: 'sneakers',
-        imageUrl: 'https://i.ibb.co/0jqHpnp/sneakers.png',
-        id: 3,
-        linkUrl: 'shop/sneakers',
-    },
-    {
-        title: 'womens',
-        imageUrl: 'https://i.ibb.co/GCCdy8t/womens.png',
-        size: 'large',
-        id: 4,
-        linkUrl: 'shop/womens',
-    },
-    {
-        title: 'mens',
-        imageUrl: 'https://i.ibb.co/R70vBrQ/men.png',
-        size: 'large',
-        id: 5,
-        linkUrl: 'shop/mens',
-    },
-];
+import { useSelector } from 'react-redux';
+import { RootState } from '../../store/reducers/reducers';
 
 export const Directory: React.FC = () => {
-    const [sections] = React.useState(sectionData);
+    const sections = useSelector((state: RootState) => state.directory.sections);
 
     return (
         <div className="directory-menu">
