+++ src/pages/shop/shop.tsx
@@ -1,14 +1,8 @@
 import * as React from 'react';
-import { CollectionPreview } from '../../components/collection-preview/collection-preview';
-import { SHOP_DATA } from './shop.constants';
+import { CollectionsOverview } from '../../components/collections-overview/collections-overview';
 
-export const ShopPage: React.FC = () => {
-    const [collections] = React.useState(SHOP_DATA);
-    return (
-        <div className="shop-page">
-            {collections.map(({ id, ...otherCollectionProps }) => (
-                <CollectionPreview key={id} {...otherCollectionProps} />
-            ))}
-        </div>
-    );
-};
+export const ShopPage: React.FC = () => (
+    <div className="shop-page">
+        <CollectionsOverview />
+    </div>
+);
